feat(TON-3119): Add console message hex in sample
