+++ backend-nodejs/@tonx-adapter/my-nft-transfer/src/index.ts
@@ -1,7 +1,6 @@
 import {
     beginCell,
     Address,
-    TonClient,
     WalletContractV4,
     internal,
     external,
@@ -90,7 +89,7 @@ async function transferNFT({
     await client.sendFile(signedTransaction);
 
     const hash = externalMessageCell.hash().toString('hex');
-    console.log('Transaction hash:', hash);
+    console.log('message hash:', hash);
     return hash;
 }
