+++ backend-nodejs/@tonx-core/my-jetton-transfer/src/index.ts
@@ -113,9 +113,9 @@ async function transferJettons(params: JettonTransferParams): Promise<TransferRe
         await provider.sendMessage(boc);
 
         const hash = externalMessageCell.hash().toString('hex');
-        console.log("Transfer successful!");
-        console.log("Transaction hash:", hash);
+        console.log("message hash:", hash);
 
+        console.log("Transfer successful!");
         return {
             success: true,
             hash
