+++ backend-nodejs/@tonx-core/my-ton-transfer/src/index.ts
@@ -92,9 +92,9 @@ async function transferTON(params: TONTransferParams): Promise<TransferResult> {
         await provider.sendMessage(boc);
 
         const hash = externalMessageCell.hash().toString('hex');
-        console.log("Transfer successful!");
-        console.log("Transaction hash:", hash);
+        console.log("message hash:", hash);
 
+        console.log("Transfer successful!");
         return {
             success: true,
             hash
