+++ react-ts-app/@tonx-core/my-jetton-transfer/src/App.tsx
@@ -105,6 +105,8 @@ export default function App() {
       await provider.sendMessage(boc);
 
       const hash = externalMessageCell.hash().toString('hex');
+      console.log("message hash: " + hash);
+
       setStatus("Transfer successful!");
 
     } catch (error) {
