+++ react-ts-app/@tonx-core/my-ton-transfer/src/App.tsx
@@ -83,6 +83,8 @@ export default function App() {
       await provider.sendMessage(boc);
 
       const hash = externalMessageCell.hash().toString('hex');
+      console.log("message hash: " + hash);
+
       setStatus("Transfer successful!");
     } catch (error) {
       console.error('Transfer failed:', error);
